LayoutTests/platform/ios-sim-deprecated/http/tests/security/dataURL/xss-DENIED-from-data-url-sub-frame-to-data-url-sub-frame-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "null" from accessing a frame with origin "null". Protocols, domains, and ports must match.
This tests that a data: URL subframe can't access a sibling data: URL subframe.

 

--------
Frame: 'firstSubFrame'
--------
Inner iframe.

Pass: Cross frame access from a sibling data: URL was denied.

--------
Frame: 'secondSubFrame'
--------
Inner iframe.

Test
